requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
alembic==1.10.2
async-timeout==4.0.2
attrs==22.2.0
bitarray==2.7.3
certifi==2021.10.8
charset-normalizer==3.1.0
click==8.1.3
cytoolz==0.12.1
eth-abi==4.0.0b2
eth-account==0.8.0
eth-hash==0.5.1
eth-keyfile==0.6.1
eth-keys==0.4.0
eth-rlp==0.3.0
eth-typing==3.3.0
eth-utils==2.1.0
Flask==2.2.3
Flask-Migrate==4.0.4
Flask-SQLAlchemy==3.0.3
frozenlist==1.3.3
greenlet==2.0.2
hexbytes==0.3.0
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
loguru==0.6.0
lru-dict==1.1.8
Mako==1.2.4
MarkupSafe==2.1.2
multidict==6.0.4
netaddr==0.8.0
parsimonious==0.9.0
pkg_resources==0.0.0
pkgutil_resolve_name==1.3.10
protobuf==4.22.1
psycopg2-binary==2.9.5
pycryptodome==3.17
pyrsistent==0.19.3
python-dotenv==1.0.0
regex==2022.10.31
requests==2.28.2
rlp==3.0.0
SQLAlchemy==2.0.6
toolz==0.12.0
typing_extensions==4.5.0
urllib3==1.26.15
varint==1.0.2
web3==6.0.0
websockets==10.4
Werkzeug==2.2.3
yarl==1.8.2
zipp==3.15.0
